[orc-rt] Add support object library, inline small system APIs
Create orc-rt/lib/support and move the support-layer implementation
files there as an object library that Bedrock embeds. This is a first
step toward the Support/Bedrock/SPIRE split, in which Support is always
an object library embedded into the libraries that ship.

Since Support will be JIT-linked into each Session, trim it down where
possible: per-system implementations that can be accessed via Bedrock
move there (Memory, DynamicLibrary, etc). Environment and Errno become
header only. CacheControl collapses into a single header.

An upcomming commit will update the unit tests to reflect this split.
