+++ orc-rt/include/orc-rt-internal/support/Environment.h
@@ -1,4 +1,4 @@
-//===- Environment.cpp - ORC-RT executor environment access ---------------===//
+//===- Environment.h - ORC-RT executor environment access -------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,16 +6,12 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// Implementation of the environment helpers declared in Environment.h.
+// Helpers for reading the process environment.
 //
 //===----------------------------------------------------------------------===//
 
-// Must precede any include so <stdlib.h> exposes glibc's secure_getenv.
-#ifndef _GNU_SOURCE
-#define _GNU_SOURCE
-#endif
-
-#include "orc-rt-internal/bedrock/Environment.h"
+#ifndef ORC_RT_INTERNAL_SUPPORT_ENVIRONMENT_H
+#define ORC_RT_INTERNAL_SUPPORT_ENVIRONMENT_H
 
 #include <stdlib.h>
 
@@ -24,9 +20,27 @@
 #include <unistd.h> // for issetugid
 #endif
 
+#if defined(__GLIBC__)
+// Declared here rather than by defining _GNU_SOURCE before <stdlib.h>: a header
+// cannot set a feature macro reliably, since <stdlib.h> may already have been
+// included by the time this one is reached. The signature matches glibc's,
+// whose __THROW expands to noexcept in C++.
+extern "C" char *secure_getenv(const char *Name) noexcept;
+#endif
+
 namespace orc_rt {
 
-const char *secureGetenv(const char *Name) {
+/// Like getenv, but returns null when the process is running with elevated
+/// privileges (e.g. a set-user-ID or set-group-ID program), so that a variable
+/// in an attacker-controlled environment cannot influence a privileged host.
+///
+/// Use this for any environment variable whose value has security-relevant
+/// effects, e.g. choosing a file to open.
+///
+/// This is header-only so that Support carries no per-system objects: Support
+/// is embedded into every library that ships, including the JIT-linked SPIRE,
+/// so anything it compiles is code those libraries have to carry.
+inline const char *secureGetenv(const char *Name) {
 #if defined(__GLIBC__)
   // secure_getenv returns null in "secure execution" mode, which the kernel
   // sets for set-user-ID / set-group-ID programs (and other privilege
@@ -50,3 +64,5 @@ const char *secureGetenv(const char *Name) {
 }
 
 } // namespace orc_rt
+
+#endif // ORC_RT_INTERNAL_SUPPORT_ENVIRONMENT_H
