+++ orc-rt/lib/CMakeLists.txt
@@ -1,1 +1,3 @@
+# Support must be configured first: Bedrock embeds its objects.
+add_subdirectory(support)
 add_subdirectory(bedrock)
