+++ orc-rt/lib/support/Logging.cpp
