+++ orc-rt/lib/support/Logging_oslog.cpp
